Split starrocks.go into one file per struct for readability
command_context.go (commandContext + plugin.Handler methods),
cluster_context.go (clusterContext, no methods), job_context.go
(jobContext + collectResults), backend_client_pool.go
(backendClientPool + fetchEndpoint/locationHost). starrocks.go keeps
just the shared consts/telemetry vars and dialFlightClient, used by
both command_context.go and backend_client_pool.go.
