+++ internal/pkg/object/command/starrocks/cluster_context.go
@@ -1,0 +1,7 @@
+package starrocks
+
+type clusterContext struct {
+	Endpoint string `yaml:"endpoint" json:"endpoint"`
+	Database string `yaml:"database,omitempty" json:"database,omitempty"`
+	UseTLS   bool   `yaml:"use_tls,omitempty" json:"use_tls,omitempty"`
+}
